refactor!: update the type of total count from uint32 to int64

+++ dtos/common/count.go
@@ -3,11 +3,11 @@ package common
 // CountResponse defines the Response Content for GET count DTO.
 type CountResponse struct {
 	BaseResponse `json:",inline"`
-	Count        uint32 `json:"count"`
+	Count        int64 `json:"count"`
 }
 
 // NewCountResponse creates new CountResponse with all fields set appropriately
-func NewCountResponse(requestId string, message string, statusCode int, count uint32) CountResponse {
+func NewCountResponse(requestId string, message string, statusCode int, count int64) CountResponse {
 	return CountResponse{
 		BaseResponse: NewBaseResponse(requestId, message, statusCode),
 		Count:        count,
